tests(btc_server): move integration tests to tests dir
